pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qualdatan-core"
version = "0.1.0"
description = "Qualdatan core library: qualitative data analysis primitives (PDF, QDPX, LLM coding, run state) without any UI or domain-specific logic."
readme = "README.md"
requires-python = ">=3.11"
license = "AGPL-3.0-only"
authors = [
    { name = "Friedrich Schroedter" },
]
keywords = ["qualitative-analysis", "maxqda", "refi-qda", "claude", "coding"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "anthropic>=0.40.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pymupdf>=1.27.0",
    # Used for progress output in the orchestrator. TUI/Desktop can still
    # wrap via qualdatan_core.events for structured consumption, but the
    # orchestrator prints human-readable status inline by default.
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.26",
]

[project.urls]
Homepage = "https://github.com/GeneralPawz/Qualdatan"
Repository = "https://github.com/Qualdatan/core"
Issues = "https://github.com/Qualdatan/core/issues"

# ---------------------------------------------------------------------------
# Packaging
# ---------------------------------------------------------------------------
# src/qualdatan_core/ is the importable package. Phase A.2 adds subpackages
# (pdf/, qdpx/, steps/, coding/, llm/, facets/) without changing the
# top-level import path.
[tool.setuptools.packages.find]
where = ["src"]
include = ["qualdatan_core*"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-q",
    "--strict-markers",
]
markers = [
    "slow: tests that make real API calls or are otherwise slow",
]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
# E501: line length handled by formatter; B008: function-call default args are fine;
# UP042: `class X(str, Enum)` is intentional — StrEnum changes __str__ semantics.
ignore = ["E501", "B008", "UP042"]

# ---------------------------------------------------------------------------
# Mypy — lauft vorerst non-blocking im CI (`continue-on-error: true`),
# damit die Baseline sichtbar wird, ohne den Build zu blockieren.
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
files = ["src/qualdatan_core"]
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
no_implicit_optional = true
# Schrittweise verschaerfen; vorerst keine `disallow_untyped_defs`.

[[tool.mypy.overrides]]
module = [
    "anthropic.*",
    "fitz.*",
    "pymupdf.*",
    "docx.*",
    "openpyxl.*",
    "yaml.*",
    "rich.*",
    "dotenv.*",
]
ignore_missing_imports = true

[tool.ruff.lint.per-file-ignores]
# Tests: unused imports are part of fixture wiring; module-level imports at
# the bottom are a documented test pattern (setup before import).
"tests/*" = ["B", "F401", "F841", "E402"]
# pdf_coder calls warnings.filterwarnings before the anthropic import, which
# is necessary to suppress a pymupdf deprecation warning at import time.
"src/qualdatan_core/pdf_coder.py" = ["E402"]
# app_db/__init__.py re-exports submodules after defining AppDB to avoid
# circular imports between AppDB and its DAOs.
"src/qualdatan_core/app_db/__init__.py" = ["E402"]
